Python 2 support, take 2
Drop the hacky Python2 messaging module. The split of messaging into smaller components per #114 and a couple of small tweaks allowed us to support Python2 for both Messaging, and the CSV and JSON modules. Header conversion still unsupported for Python 2, and we still have no in intent on supporting Python 2 officially moving forward.

Note that also renamed msgtools/lib/json and csv to be prefixed with 'msg'. This prevents collisions with the built-in Python libraries of the same name. It is technically possible to re-use built in names but the Python importer isn't guaranteed to be feature stable in future releases with respect a fix we put in now. Renaming to avoid the collision is less fragile.